+++ keywords/webui/data_protection.py
@@ -1,4 +1,6 @@
+import xpaths
 from keywords.webui.common import Common as COM
+from keywords.webui.navigation import Navigation as NAV
 
 
 class Data_Protection:
@@ -25,4 +27,18 @@ class Data_Protection:
             COM.click_on_element('//*[starts-with(@data-test,"button-snapshot") and contains(@data-test,"-delete-row-action")]')
             COM.assert_confirm_dialog()
         assert COM.is_visible('//*[starts-with(@data-test,"button-snapshot") and contains(@data-test,"-delete-row-action")]') is False
+
+    @classmethod
+    def delete_all_snapshots(cls):
+        if COM.assert_page_header('Snapshots') is False:
+            NAV.navigate_to_periodic_snapshots()
+        COM.set_checkbox('column-select-all')
+        if COM.is_visible(xpaths.common_xpaths.button_field('delete-selected')) is True:
+            COM.click_button('delete-selected')
+            COM.set_checkbox('confirm')
+            COM.click_button('delete')
+            # delay to allow delete to complete
+            assert COM.is_visible(xpaths.common_xpaths.button_field('close')) is True
+            COM.click_button('close')
+            COM.assert_text_is_visible('No records have been added yet')
